optimize diff computation
Previously the diff of the commit and its parent was computed each time
for every path being checked. A recent change (performance regression?)
slowed that down considerably.

This optimization instead adds each path to the `DiffOptions` and the
diff is only computed once. Then for each path it checks if the Pathspec
of the file is present in the Diff using Pathspec.match_diff.

This provided a big performance boost so that speeds are nearly back to
what they were before.

This also refactors the code to be simpler and more concise.